CONTENT BOX
`border-box` includes borders and paddings in the calculation. This is
very convenient in most cases, but the half paddings/margins we're
applying to get the borders right mean that there is no guarantee the
end result will respect the grid. `content-box` on the other hand can
give us that guarantee -- and since _we_ give the guarantee that the
borders line up by construction, the end result should be more robust to
misalignment (if slightly less convenient to reason about).